Move header search box to the separate component
