+++ src/components/Header/Header.js
@@ -2,9 +2,7 @@ import PropTypes from "prop-types"
 import styled from "styled-components"
 
 import { useState } from "react"
-import { useHistory } from "react-router-dom"
 
-// import { Content } from "../shared/Containers"
 import Logo from "./Logo"
 
 import Navigation from "./Navigation"
@@ -19,15 +17,7 @@ import {
 } from "./MobileMenu"
 // import MainGridSharedStyle from "../../pages/sharedStyles"
 import { MainGrid } from "../shared/Containers"
-
-import { ReactComponent as SearchIcon } from "../../icons/icon-search.svg"
-import { ReactComponent as ArrowIcon } from "../../icons/search-arrow-icon.svg"
-import {
-  ButtonArrow,
-  SearchForm,
-  TextInput,
-  ContainerSearchIcon,
-} from "../shared/SearchBoxModules"
+import SearchBox from "./SearchBox"
 
 const StyledHeader = styled.header`
   padding: 0.5em 0 0.5em;
@@ -65,33 +55,6 @@ function Header({ noSearchBox }) {
     }
   }
 
-  const [searchQuery, setSearchQuery] = useState("")
-
-  const handleChange = (event) => {
-    setSearchQuery(event.target.value)
-  }
-
-  const history = useHistory()
-
-  const handleSearchSubmit = (event) => {
-    event.preventDefault()
-    console.log("search submitted")
-    event.stopPropagation()
-
-    // clean up search query
-    const cleanedSearchQuery = searchQuery
-      // make search query lowercase
-      .toLowerCase()
-      // remove white spaces from the start and from the end of a string
-      .trim()
-      // replace multiple spaces with a single space
-      .replace(/\s\s+/g, " ")
-
-    history.push(`/search/${cleanedSearchQuery}`)
-    // reset input in search box
-    setSearchQuery("")
-  }
-
   return (
     <>
       <StyledHeader>
@@ -100,25 +63,7 @@ function Header({ noSearchBox }) {
             <ContainerFlex showHamburgerMenu={showHamburgerMenu}>
               <Logo />
               {/* hide search box if Header receieved 'noSearchBox' prop or if it is a mobile layout */}
-              {isDesktopResolution && showSearchBox && (
-                <SearchForm onSubmit={handleSearchSubmit}>
-                  <ContainerSearchIcon>
-                    <SearchIcon onClick={handleSearchSubmit} />
-                  </ContainerSearchIcon>
-                  <TextInput
-                    placeholder="Search music..."
-                    value={searchQuery}
-                    onChange={handleChange}
-                  />
-                  <ButtonArrow
-                    type="submit"
-                    isEmpty={searchQuery}
-                    onClick={handleSearchSubmit}
-                  >
-                    <ArrowIcon />
-                  </ButtonArrow>
-                </SearchForm>
-              )}
+              {isDesktopResolution && showSearchBox && <SearchBox />}
             </ContainerFlex>
 
             {isTabletOrMobile ? (
